HH-218702 move healthcheck interfaces to nab-jdbc-common module
