+++ nab-data-source/src/main/java/ru/hh/nab/datasource/healthcheck/HealthCheckResultImpl.java
@@ -1,4 +1,6 @@
 package ru.hh.nab.datasource.healthcheck;
 
+import ru.hh.nab.jdbc.common.healthcheck.HealthCheckResult;
+
 public record HealthCheckResultImpl(boolean healthy) implements HealthCheckResult {
 }
