+++ nab-hibernate-persistence-provider/src/main/java/ru/hh/nab/hibernate/datasource/RoutingDataSource.java
@@ -15,11 +15,11 @@ import org.springframework.lang.Nullable;
 import ru.hh.nab.datasource.DataSourceContextUnsafe;
 import static ru.hh.nab.datasource.DataSourceSettings.DATASOURCE_NAME_FORMAT;
 import ru.hh.nab.datasource.DataSourceType;
-import ru.hh.nab.datasource.healthcheck.HealthCheck;
-import ru.hh.nab.datasource.healthcheck.HealthCheckDataSource;
 import ru.hh.nab.jdbc.common.DataSourcePropertiesStorage;
 import ru.hh.nab.jdbc.common.datasource.NamedDataSource;
 import ru.hh.nab.jdbc.common.ext.JdbcExtension;
+import ru.hh.nab.jdbc.common.healthcheck.HealthCheck;
+import ru.hh.nab.jdbc.common.healthcheck.HealthCheckDataSource;
 import ru.hh.nab.metrics.Counters;
 import ru.hh.nab.metrics.StatsDSender;
 import ru.hh.nab.metrics.Tag;
