+++ nab-jdbc-common/src/main/java/ru/hh/nab/jdbc/common/healthcheck/HealthCheckDataSource.java
@@ -1,4 +1,4 @@
-package ru.hh.nab.datasource.healthcheck;
+package ru.hh.nab.jdbc.common.healthcheck;
 
 public interface HealthCheckDataSource {
   String getDataSourceName();
