+++ nab-jdbc-common/src/main/java/ru/hh/nab/jdbc/common/healthcheck/HealthCheckResult.java
@@ -1,4 +1,4 @@
-package ru.hh.nab.datasource.healthcheck;
+package ru.hh.nab.jdbc.common.healthcheck;
 
 public interface HealthCheckResult {
   boolean healthy();
